# docs/backend/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: AI Food Ordering Internal API
  version: 1.0.0
  description: |
    Internal REST API for the AI Food Ordering backend.

    The API is intended for trusted internal clients such as the MCP service
    and Telegram integration. All API routes require `X-Internal-Api-Token`.
    Session-bound routes additionally require `X-Session-Token`.

    The backend owns restaurant selection, cart state, prices, totals, order
    state and Dots integration. Clients must not attempt to submit internal
    identifiers or calculated totals.

    Automated API/E2E tests run against the Laravel application and mock the
    external Dots HTTP boundary where appropriate.
servers:
  - url: http://localhost:8080
    description: Local Docker environment
...
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create an order from the current cart
      description: |
        Checks out the current cart and creates an order.

        The MVP checkout is fixed to pickup and cash. The client provides only
        `delivery_time`; restaurant, Dots company/address/city identifiers,
        receiving type, payment type, customer data, cart items, totals and
        currency are resolved by the backend.

        Before submitting an order, the backend validates cart prices through
        Dots. The Dots `totalPrice` is authoritative and is persisted as the
        order total, so the final order total may differ from the local catalog
        subtotal because of promotions or Dots-side pricing rules.

        `Idempotency-Key` is mandatory. Repeating the same request with the
        same key and session returns the existing order with `200 OK` and does
        not submit another order to Dots. A new order returns `201 Created`.

        Dots may accept creation asynchronously. A newly accepted order may
        initially have status `creating`; `GET /api/orders/current` later
        refreshes it to `created` when Dots confirms the order.
      security:
        - InternalApiToken: []
          SessionToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              asap:
                summary: ASAP pickup
                value:
                  delivery_time: 0
              scheduled:
                summary: Scheduled pickup using Unix timestamp
                value:
                  delivery_time: 1786442400
      responses:
        '200':
          description: Existing order returned for the same idempotency key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '201':
          description: New local order created and submitted/accepted for creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Active restaurant or current cart was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            Checkout conflict, for example an inactive cart, changed cart,
            another order already created for the cart, or an idempotency key
            already belonging to another session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Validation or checkout rejection. Examples include an invalid
            request, expired/empty cart, unavailable product, missing/invalid
            customer contact, or a definite 4xx rejection from Dots.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Dots returned a server error or invalid response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Connection to the ordering service failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    InternalApiToken:
      type: apiKey
      in: header
      name: X-Internal-Api-Token
      description: Shared secret used by trusted internal services.
    SessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        64-character lowercase hexadecimal token returned by
        `POST /api/sessions`. The session must exist, be active and unexpired.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated idempotency key for order creation. Maximum length is
        128 characters. Reusing the same key in the same session returns the
        existing order without a second Dots create request.
      schema:
        type: string
        minLength: 1
        maxLength: 128
      example: order-01J5X3MYE73ZZ7C6CPV4RDQBR2
  responses:
    Unauthorized:
      description: Missing, malformed, expired or invalid internal/session authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Unauthenticated.
...
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      additionalProperties: true
      properties:
        message:
          type: string
      example:
        message: Unauthenticated.
    ValidationErrorResponse:
      type: object
      required: [message, errors]
      additionalProperties: true
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            quantity:
              - The quantity field must be at least 1.
    Money:
      type: string
      pattern: '^-?\d+\.\d{2}$'
      description: Decimal monetary amount serialized as a string with two fractional digits.
      examples: ['100.00', '80.00']
...
    CreateOrderRequest:
      type: object
      required: [delivery_time]
      additionalProperties: true
      properties:
        delivery_time:
          type: integer
          minimum: 0
          description: |
            `0` means ASAP in Dots. A positive value is a future Unix timestamp
            used to schedule pickup.
          example: 0
      description: |
        MVP checkout is fixed to pickup + cash. The following internal fields
        must be absent: `idempotency_key`, restaurant/company/address/city
        identifiers, receiving/delivery/payment types, customer, items, total
        and currency.
    OrderItem:
      type: object
      required:
        - product_id
        - external_product_id
        - name
        - quantity
        - unit_price
        - total
      properties:
        product_id:
          type: [integer, 'null']
          minimum: 1
          description: Local product ID snapshot; may become null if the local product is deleted later.
        external_product_id:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
    Order:
      type: object
      required:
        - id
        - external_order_id
        - status
        - failure_message
        - receiving_type
        - total
        - currency
        - items
      properties:
        id:
          type: integer
          minimum: 1
        external_order_id:
          type: [string, 'null']
          description: Dots order ID when Dots has returned one.
        status:
          type: string
          enum:
            - draft
            - creating
            - created
            - awaiting_payment
            - paid
            - failed
            - cancelled
        failure_message:
          type: [string, 'null']
          description: Last known creation failure/uncertainty information when present.
        receiving_type:
          type: string
          enum: [pickup]
        total:
          $ref: '#/components/schemas/Money'
          description: Final Dots-validated order total.
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: UAH
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Order'